Added Data Initialization
Aim: This commit is intended to add initial recipe data bootstrapping to
the application.

Change: This commit added the recipeBootstrap class, which would
generate two boilerplate recipes each time when the Application Context
refreshes. It also modifed the index controller in order to display all
recipes at the index page, and added the recipeService service layer to
manage recipes in the application.