Added Recipe Update Handler
ABOUT: This change aimed to update the recipe update logic in the recipe
controller. The update shares the recipe form view with create.

CHANGE: This change modified the index view to include entry points to
the recipe update, the recipe controller to include relevant route
mappings, and added relevant recipe controller tests.